requirements.txt
Bottleneck 
certifi==2021.10.8
cycler 
fonttools==4.25.0
joblib 
kiwisolver 
matplotlib 
mkl-fft==1.3.1
mkl-random 
mkl-service==2.4.0
munkres==1.1.4
numexpr 
numpy 
packaging
pandas==1.4.1
Pillow==9.0.1
pyparsing 
python-dateutil 
pytz==2021.3
ray
scikit-learn 
scipy 
sip
six 
threadpoolctl 
tornado 
